Create README.md
This is an exemple of a simple single screen app that will facilitate customers to have the most important information about a company on screen at the push of a thumb. 
It empowers customers with information about the mission of a company while adding connectivity by e-mail or phone and showing the geographical company location and route on a map.